Replaced my Vec3D with LWJGL's Vector3f.

+++ src/main/java/jigglyslimes/JigglySlimes.java
@@ -13,9 +13,9 @@ import org.apache.logging.log4j.Logger;
 public class JigglySlimes {
 
     public static final String MODID = "jigglyslimes";
-    public static final double GRAVITY = -32.0; // In m/s^2
-    public static final double AIR_FRICTION = 50.0;
-    public static final double AIR_DENSITY = 1.2; // In kg/m^3
+    public static final float GRAVITY = -32.0F; // In m/s^2
+    public static final float AIR_FRICTION = 50.0F;
+    public static final float AIR_DENSITY = 1.2F; // In kg/m^3
 
     public static final Logger LOGGER = LogManager.getLogger();
